Update App Name and Localization Strings
Refactor the application name to use localized strings for better internationalization. This change includes updating the AndroidManifest.xml to reference the app name from string resources, adding string resource files for both Arabic and English, and updating the corresponding ARB files for localization. Additionally, fixed an issue in the sura_details_screen.dart file where the color property for an image asset was commented out, causing it to not respect the dark mode setting.