Move UpgradeWrapper to shared components and update references
The `UpgradeWrapper` widget was previously defined in `home_screen.dart`, which is not ideal for reusability. This commit moves the `UpgradeWrapper` widget to the `shared/components/upgrade_wrapper.dart` file and updates the `export_manager.dart` to include the new part. Additionally, the `_onAzkarButtonPressed` method in `sebha_screen.dart` was commented out and replaced with `_onRotateTap` to reflect the new button functionality.